refactor(authenticate): validate moved to parent class
