+++ modules/authentication/src/constants/TokenType.ts
@@ -5,3 +5,4 @@ export enum TokenType {
   TWO_FA_VERIFICATION_TOKEN = 'TWO_FA_VERIFICATION_TOKEN',
   VERIFY_PHONE_NUMBER_TOKEN = 'VERIFY_PHONE_NUMBER_TOKEN',
 }
+
