+++ modules/authentication/src/handlers/facebook/facebook.ts
@@ -1,7 +1,6 @@
 import { isNil } from 'lodash';
-import ConduitGrpcSdk, { ConduitError, GrpcError } from '@conduitplatform/conduit-grpc-sdk';
+import ConduitGrpcSdk, { GrpcError } from '@conduitplatform/conduit-grpc-sdk';
 import { status } from '@grpc/grpc-js';
-import { ConfigController } from '../../config/Config.controller';
 import axios, { AxiosRequestConfig } from 'axios';
 import { Payload } from '../AuthenticationProviders/interfaces/Payload';
 import { OAuth2 } from '../AuthenticationProviders/OAuth2';
@@ -9,28 +8,10 @@ import { FacebookSettings } from './facebook.settings';
 import { FacebookUser } from './facebook.user';
 
 export class FacebookHandlers extends OAuth2<Payload, FacebookSettings> {
-  private initialized: boolean = false;
 
   constructor(grpcSdk: ConduitGrpcSdk, settings: FacebookSettings) {
     super(grpcSdk, 'facebook', settings);
   }
-
-  async validate(): Promise<Boolean> {
-    const authConfig = ConfigController.getInstance().config;
-
-    if (!authConfig.facebook.enabled) {
-      console.log('Facebook not active');
-      throw ConduitError.forbidden('Facebook auth is deactivated');
-    }
-    if (!authConfig.facebook.clientId) {
-      console.log('Facebook not active');
-      throw ConduitError.forbidden('Cannot enable facebook auth due to missing clientId');
-    }
-    console.log('Facebook is active');
-    this.initialized = true;
-    return true;
-  }
-
 
   async connectWithProvider(details: { accessToken: string, clientId: string, scope: string }): Promise<FacebookUser> {
     if (!this.initialized)
