+++ modules/authentication/src/handlers/figma/figma.ts
@@ -1,7 +1,6 @@
 import { isNil } from 'lodash';
-import ConduitGrpcSdk, { ConduitError, GrpcError } from '@conduitplatform/conduit-grpc-sdk';
+import ConduitGrpcSdk, { GrpcError } from '@conduitplatform/conduit-grpc-sdk';
 import { status } from '@grpc/grpc-js';
-import { ConfigController } from '../../config/Config.controller';
 import axios, { AxiosRequestConfig } from 'axios';
 import { OAuth2 } from '../AuthenticationProviders/OAuth2';
 import { FigmaUser } from './figma.user';
@@ -9,28 +8,10 @@ import { FigmaSettings } from './figma.settings';
 import { SlackUser } from '../slack/slack.user';
 
 export class FigmaHandlers extends OAuth2<FigmaUser, FigmaSettings> {
-  private initialized: boolean = false;
 
   constructor(grpcSdk: ConduitGrpcSdk, settings: FigmaSettings) {
     super(grpcSdk, 'figma', settings);
   }
-
-  async validate(): Promise<Boolean> {
-    const authConfig = ConfigController.getInstance().config;
-
-    if (!authConfig.figma.enabled) {
-      console.log('Figma not active');
-      throw ConduitError.forbidden('Figma auth is deactivated');
-    }
-    if (!authConfig.facebook.clientId) {
-      console.log('Figma not active');
-      throw ConduitError.forbidden('Cannot enable figma auth due to missing clientId');
-    }
-    console.log('Figma is active');
-    this.initialized = true;
-    return true;
-  }
-
 
   async connectWithProvider(details: { accessToken: string, clientId: string, scope: string }): Promise<FigmaUser> {
     if (!this.initialized)
