+++ modules/authentication/src/handlers/twitch/twitch.ts
@@ -1,36 +1,13 @@
-import ConduitGrpcSdk, { ConduitError } from '@conduitplatform/conduit-grpc-sdk';
+import ConduitGrpcSdk from '@conduitplatform/conduit-grpc-sdk';
 import axios from 'axios';
-import { ConfigController } from '../../config/Config.controller';
 import { TwitchUser } from './twitch.user';
 import { TwitchSettings } from './twitch.settings';
 import { OAuth2 } from '../AuthenticationProviders/OAuth2';
 
 export class TwitchHandlers extends OAuth2<TwitchUser, TwitchSettings> {
-  private initialized: boolean = false;
 
   constructor(grpcSdk: ConduitGrpcSdk, settings: TwitchSettings) {
     super(grpcSdk, 'twitch', settings);
-  }
-
-  async validate(): Promise<Boolean> {
-    const authConfig = ConfigController.getInstance().config;
-    if (!authConfig.twitch.enabled) {
-      console.log('twitch not active');
-      throw ConduitError.forbidden('Twitch auth is deactivated');
-    }
-    if (
-      !authConfig.twitch ||
-      !authConfig.twitch.clientId ||
-      !authConfig.twitch.clientSecret
-    ) {
-      console.log('twitch not active');
-      throw ConduitError.forbidden(
-        'Cannot enable twitch auth due to missing clientId or client secret',
-      );
-    }
-    console.log('twitch is active');
-    this.initialized = true;
-    return true;
   }
 
   async connectWithProvider(details: { accessToken: string, clientId: string, scope: string }): Promise<TwitchUser> {
